style: Update gif layout to css grid
Initially styled the grif layout with flexbox, but I wanted it to mimic the giphy grid where vertical space was filled, leaving no awkward gaps between items. Switched the layout to grid to achieve this look. Also updated error messaging to disappear when the user corrects the error (e.g. enters in a new query).