style: Add link styling on gifs
Added a link icon that appears on hover that directs users to the Giphy site to interact further with the gif. On desktop it appears on hover and is accessible. On mobile it appears in the lefthand corner of each gif. Styling on mobile should be designed better, but tabling for now.